requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Data acquisition
requests>=2.31.0
tqdm>=4.66.0

# Google BigQuery
google-cloud-bigquery>=3.11.0
google-auth>=2.22.0
db-dtypes>=1.1.1

# Configuration
pyyaml>=6.0

# Excel export
openpyxl>=3.1.0

# Geographic data (optional)
geopandas>=0.13.0
shapely>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
ipython>=8.14.0
jupyter>=1.0.0
black>=23.0.0
